+++ testar/src/nl/ou/testar/ReinforcementLearning/Policies/OptimisticQValuesInitializationPolicy.java
@@ -9,9 +9,9 @@ import java.util.Set;
 
 public class OptimisticQValuesInitializationPolicy implements Policy {
     private final GreedyPolicy greedyPolicy;
-    private final double maxQValue;
+    private final float maxQValue;
 
-    public OptimisticQValuesInitializationPolicy(final GreedyPolicy greedyPolicy, final double maxQValue) {
+    public OptimisticQValuesInitializationPolicy(final GreedyPolicy greedyPolicy, final float maxQValue) {
         this.greedyPolicy = greedyPolicy;
         this.maxQValue = maxQValue;
     }
